refactor: useLocaleText
