+++ packages/bui-core/src/locales/hooks/useLocaleText.ts
@@ -1,5 +1,5 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { ThemeProps, CN, useTheme } from '@bifrostui/react';
+import { CN } from '../index';
+import { ThemeProps, useTheme } from '../../ThemeProvider';
 
 // 官网默认语言是zh-CN 外部使用若不传locale，则默认显示中文
 function useLocaleText(componentName: string, theme?: Partial<ThemeProps>) {
