+++ packages/bui-core/src/locales/index.ts
@@ -1,4 +1,5 @@
 export { default as EN } from './en-US';
 export { default as CN } from './zh-CN';
 export { default as TW } from './zh-TW';
+export { default as useLocaleText } from './hooks/useLocaleText';
 export * from './base';
